waylandsink: Add use-subsurface option as NOP option
The main purpose of this option was to scanout the image
when the application enables the fullscreen mode.

But now this option is not necessary because scanout
can be done without this option.
When this option was made weston didn't support setting opaque
for subbsurface, but since weston v2.0, it is supported.

For backward comaptibility, add this option as NOP
and deprecated option.

Porting from 1.18.5 to 1.22.11
Signed-off-by: duytruong <duy.truong.yw@renesas.com>